pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "praxis"
version = "0.1.0"
description = "Policy-driven AI workflow governance"
authors = ["jayers"]
readme = "README.md"
packages = [{include = "praxis", from = "src"}]
include = ["src/praxis/templates/**"]

[tool.poetry.dependencies]
python = ">=3.10"
typer = ">=0.9.0"
pydantic = ">=2.0"
pyyaml = ">=6.0"
questionary = ">=2.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-bdd = "^8.0"
ruff = "^0.8"
mypy = "^1.0"
types-pyyaml = "^6.0"
pre-commit = "^4.5.1"

[tool.poetry.scripts]
praxis = "praxis.cli:app"

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["E501"]  # Allow long lines in test fixtures

[tool.mypy]
python_version = "3.10"
strict = true
packages = ["praxis"]

[[tool.mypy.overrides]]
module = ["yaml"]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
